requirements.txt
appdirs==1.4.4
attrs==20.2.0
Babel==2.8.0
beautifulsoup4==4.9.1
certifi==2020.6.20
chardet==3.0.4
coverage==5.0.3
distlib==0.3.1
Django==2.2.16
django-debug-toolbar==2.2
django-extensions==2.2.8
django-extra-views==0.11.0
django-haystack==2.8.1
django-localflavor==3.0.1
django-oscar==2.0.4
-e git://github.com/django-oscar/django-oscar-paypal.git@b6eb5e5ed4c12bea6239a58f76b7c944258c32b5#egg=django_oscar_paypal
django-phonenumber-field==2.0.1
django-tables2==1.21.2
django-treebeard==4.3.1
django-webtest==1.9.7
django-widget-tweaks==1.4.5
factory-boy==2.12.0
Faker==4.1.2
filelock==3.0.12
flake8==3.8.3
idna==2.10
iniconfig==1.0.1
isort==5.5.2
mccabe==0.6.1
more-itertools==8.5.0
packaging==20.4
phonenumbers==8.12.9
Pillow==7.2.0
pluggy==0.13.1
purl==1.5
py==1.9.0
pycodestyle==2.6.0
pyflakes==2.2.0
pyparsing==2.4.7
pytest==6.0.2
pytest-cov==2.8.1
pytest-django==3.8.0
python-dateutil==2.8.1
python-stdnum==1.14
pytz==2020.1
requests==2.24.0
six==1.15.0
sorl-thumbnail==12.6.3
soupsieve==2.0.1
sqlparse==0.3.1
text-unidecode==1.3
toml==0.10.1
tox==3.14.5
urllib3==1.25.10
virtualenv==20.0.31
waitress==1.4.4
WebOb==1.8.6
WebTest==2.0.35
